refactor(project): rename project node endpoint
