+++ components/projects/ProjectProposalTable.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import { getProposals } from "~/composables/useAPIFetch";
+import { getProjectNodes } from "~/composables/useAPIFetch";
 import ApproveRejectButtons from "~/components/projects/ApproveRejectButtons.vue";
 import { formatDataRow } from "~/utils/format-data-row";
 import TableRowMetadata from "~/components/TableRowMetadata.vue";
@@ -18,7 +18,7 @@ const expandRowEntries = [];
 
 const approvalStatuses = Object.values(ApprovalStatus);
 
-const { data: response, status, error, refresh } = await getProposals();
+const { data: response, status, error, refresh } = await getProjectNodes();
 
 function parseData() {
   if (status.value === "success") {
